Revise: do the allocation of tags differently
Reuse tags by enqueuing tags as they are finished with so as to be able to
be available if all previously unused ones get used. This drops the number
of tags down to just the capital letters of the alphabet - but this will
support 26 profiles running at a time and to be perfectly honest - that
seems more than enough. Should some weirdo try and run more than that then
the additional profiles will be identified with a `[?]`...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>