BugFix: prevent crashes when clicking on empty maps
It prevents several mouse events for the mappers (both 2 and 3D) from doing
much if the map is non-existent! This should close Issue #4821.

It also:
* adds a `(bool) TRoomDB::isEmpty()` function to provide a consistent
indication that there is no rooms in the current map. (This saves having
to count all the rooms in the map {to get the `size()`} and checking the
value is not zero - technically it ought to be a tiny bit faster!)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>