BugFix: do not squash dockable mapper when visibility is toggled
The existing code was showing and hiding both the `dlgMapper` and, when it
is contained within a floating/dockable QDockWidget, that container widget.
This has the unfortunate side effect of causing that type of mapper to
shrink to a minimal size when shown a second time.

It was confusing that the code to show and hide the mapper was actually
contained within a method called `createMapper` - so that has been renamed
to `toggleVisibilityOfOrCreateMapper` and the functionality - clearly
dividable into two parts - has been refactored into two new methods that
does one part; coincidentally one of them can now be called `createMapper`!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>