BugFix: do not squash dockable mapper when visibility is toggled (#4896)
The existing code was showing and hiding both the `dlgMapper` and, when it
is contained within a floating/dockable QDockWidget, that container widget.
This has the unfortunate side effect of causing that type of mapper to
shrink to a minimal size when shown a second time.

It was confusing that the code to show and hide the mapper was actually
contained within a method called `createMapper` - so that has been renamed
to `showHideOrCreateMapper` and the functionality - clearly
dividable into two parts - has been refactored into two new methods that
does one part; coincidentally one of them can now be called `createMapper`!

Also add a `(bool) isFloatAndDockable()` method to the `dlgMapper` class
to refactor out the means to identify that instance of the class being
contained in a `QDockWidget`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
(cherry picked from commit 5e4b6b559c9dd9446f64dec7e5e1045b9d7fa532)