Fix colored buttons on macOS and Windows (#4846)
Since somewhere after Qt 5.12 the style code for the "windowssvista" style
(only) paints `QPushButton`s with native painting code that hides the
background colour set in a stylesheet setting entirely with the border.

This detail is hidden away on:
https://doc.qt.io/qt-5/stylesheet-reference.html
under QPushButton, where it says:
"Warning: If you only set a background-color on a QPushButton, the
background may not appear unless you set the border property to some
value. This is because, by default, the QPushButton draws a native border
which completely overlaps the background-color. For example,

QPushButton { background-color: red; border: none; }"

This documentation page does seems to show this from at least back to Qt
5.5 - but it does look that we have not been encountering it - certainly
neither the "Fusion" or "Windows" (old style possibly dating back to Win98
era) seem to display this behaviour - which is why this bug fix is labelled
as **Windows**.

The fix implemented here detects if the style factory being used is the
problematic "windowsvista" one (default on Windows nowadays), or something
else by appending a `--style=Windows` or `--style=Fusion` to the
`Mudlet.exe` command line invokation (possilbly with a short-cut) or does
the same thing by setting the `QT_STYLE_OVERRIDE` environmental variable to
one of those two styles; if this is the case then a flag is set and that is
used to modify some style steet strings to work around the issue.

This ought to be documented in the Lua API functions that might be used to
apply stylesheets to pushbuttons as any that set a `background-color`
without also doing something with `border` may not behave as the user
expects!

By making the style sheet creation template strings in the `mudlet`
singleton class we can generate them centrally when the application
starts and provide them globally to all the places where they are needed;
this makes the overall set of changes much smaller!

Thanks to Discord user `mars#8093` (a.k.a. `The Lost`) for demonstrating
that MacOs seems to have the same problem in it's platform specific
and default `macintosh` style as well - so this fix is also applied for that
one to.

Revised by making the style factory matching case insensitive

Simplify the debug output now that the issue has been solved but leave a
single line to debug output whether we detect a problematic style factory
or not - it will help us to debug what is going on if end-users have problems
with style (factories) and ask us to diagnose things remotely.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

Co-authored-by: Vadim Peretokin <vperetokin@gmail.com>