Bug fix make drag and drop trigger chains work again (#4901)
It seems the replacement for obsolete code I made in #4461 was
incomplete as I had missed off a defaulted but in our case required
argument to:
`[pure virtual] (QModelIndex) QAbstractItemModel::index(int, int,
                          const QModelIndex &parent = QModelIndex()) const`
in two places.

Thanks to @Edru2 for pointing this detail out and thus finding the fix!

Also polish this by adding some comments and eliminate several uses
of `goto` and as there is a chain of `if`s which each check one boolean
but only one of them should be true at a time it is worthwhile to make
all but the first of them `else if`s instead.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
(cherry picked from commit 9bd7fd88a2fd73eb321ae0e949c21310c39886c3)