Enhance: provide option to not pad copied text
This is a solution to the issue raised recently in the Mudlet Discord Help
channel which resulted in the creation of the issue that this closes #4956.

The option is presented as a context menu item on any `TConsole` window
which is checkable (and defaults to checked to mimic previous behaviour),
unchecking it will remove the padding from the first line of a multi-line
copy. However if time stamps are shown for the particular console the
padding is forced to be on and the context menu option is disabled.
Tooltips for both situations are provided.

The option is not saved between sessions because it is individual to each
console and as it is so easy to change it is simpler to not try and save it
for each and every TConsole instance.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>