Extend: also add control over timestamps being included
Following discussions on Discord the default has been switch to NOT pad
multi-line copies with spaces. The additional option to include timestamps
in the copy defaults to on for each TConsole but is not present in the
context menu until timestamps are showing in the `TConsole` - this is
because only the Main and the Editor Errors for each profile and the
Central Debug console currently display timestamps.

As the previous commit the options are not saved in the profile settings
because it is a per TConsole option and the extra code to save them is, I
feel not worth it when the option is so visible in the context menu.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>