Enhance: allow profiles to be reordered when multi-playing
This implements a requested feature to allow the tabs on the tab bar to
reordered and the profiles displayed underneath when multi-playing will
follow suite. This not being done previously was why I had to lock the
tab bar because at that time the tabs could move but the profiles did not!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>