Revise: remove static const TMainConsole::scmProperty_HostName
Given that we only use *this* instance of the `const char "HostNme"` in
THREE places in THREE separate files it was felt not to be worthwhile
abstracting it to an explicit `const` value and just to use the raw literal
where needed. As it happens the same text is used elsewhere for the same
purpose (identifying the related profile in a Qt rather than a Mudlet class
instance) there might be a more optimum way to do this in a common
universally used (maybe even pre-compiled) header file.

However that can be left for a separate PR if it is thought worthwhile...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>