Revise: try and reduce comments in PR
I'd like to leave some of them in place because there is some aspects that
are not blindingly obvious to the casual view. Also there is some code who
presence is not entirely clear to be needed to me - until someone can work
if it has to be there then it is probably okay to leave notes about what I
have worked out around it for whoever wants to do the further
investigation.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>