Enhance: allow profiles to be reordered when multi-playing (#4976)
This implements a requested feature to allow the tabs on the tab bar to
reordered and the profiles displayed underneath when multi-playing will
follow suite. This not being done previously was why I had to lock the
tab bar because at that time the tabs could move but the profiles did not!

Also remove unused function that is impacted by this PR - I knew there
was a reason why I had removed it - but neglected to include that in the
original commit.

Also update some of my Copyright lines.

Revised to remove `static const TMainConsole::scmProperty_HostName`
introduced in an earlier commit for this PR:

Given that we only use *this* instance of the `const char "HostNme"` in
THREE places in THREE separate files it was felt not to be worthwhile
abstracting it to an explicit `const` value and just to use the raw literal
where needed. As it happens the same text is used elsewhere for the same
purpose (identifying the related profile in a Qt rather than a Mudlet class
instance) there might be a more optimum way to do this in a common
universally used (maybe even pre-compiled) header file.

However that can be left for a separate PR if it is thought worthwhile...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Vadim Peretokin <vperetokin@gmail.com>