BugFix: S. Exits not staying locked when loaded from current file formats
A coding error meant the current file formats that retained the previous
storage of the special exit lock status as a `1` or `0` prefix on the
name/command for the special exit was not being corrected inserted into the
recently added separate container to store the lock status of special
exits - this meant they were effectively lost.

This will close #4999.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>