Revise: include JSON maps files as ones to try and load
Had to add something so that the (error) texts returned by
`TMap::readJsonMapFile(...)` are translated when used as part of error
messages on the main console.

It does mean that the user could abort the import of a (large) JSON map
file and then it will just try and load the next oldest one. This might
need further work if this proves objectionable.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>