Enhance: improve User Experience for encoding problems
Part 1: put in some warning icons to the left of the main command line that
can be shown to indicate some problems.

This commit handles outgoing encoder problems in that it has a clearer
error message and the icon that is shown can be clicked upon and currently
has a menu with three options:
* reset the warning - for when the user tries to send an un-encodable
character and recognises it and will learn not to use it in what they try
to send again. E.g. they try to send an emoji but the Game Server is only
set for ASCII or an extended ASCII encoding.
* ignore the warning - tells Mudlet to reset the icon/warning and not to
bother with it again in the session - it will still removed the graphemes
that it cannot encoded but will not moan about it.
* open the preferences and go straight to the control (the Game Servers
Encoding `QComboBox`) that can be changed to fix the issue - assuming the
Server can handle it...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>