BugFix: improve handling of non-map and too new files in profile's map dir (#5071)
This is an **alternative** to #5022 - and *one* of them will be needed so
that #5016 is safe - though, TBH I do not think that the latter should be
dumping JSON format files into a folder which we had previously reserved
for binary map save files!

This will also close #4275 by checking for the map version number (the
first four bytes of the file) being between 1 and 127 and rejecting it if
that test is not met.

Revised to keep GHA Clang-tidy appeased by initialising some variables
that are immediately assigned a value in the next lines following them!

Revised to include JSON maps files as ones to try and load, had to add
something so that the (error) texts returned by
`TMap::readJsonMapFile(...)` are translated when used as part of error
messages on the main console.

It does mean that the user can abort the import of a (large) JSON map
file. Personally I'd have preferred to keep the UI more uniform for both
binary and JSON map file loading but it seems from the peer-review
that it was preferred to be able to abort the load. This will, I think, trigger
some no-map code if the user does abort the loading of a JSON file...

Revised to prevent double auditing of map when loaded from JSON file

Revised to remove iteration though potential map files so reverted to
only consider the *newest* file that *could* be a map file.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>