BugFix:_restore editor auto-completer to wait for several characters
Rick (upstream maintainer of edbee-lib) adopted our idea of making the
editor auto-completer not react until a minimum number of characters has
been entered. However he, not unreasonably made it a configurable option.

In doing so he obviously removed our hard coded value of 3 but since we
were not aware of this the setting was being left at the default of 0. This
commit puts our wanted setting back in.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>