Coverity: fix issue registered as CID 1503729
It seems like there was no check for a valid TArea pointer in
`TLuaInterpreter::getMapLabel(...)` - in fact it also means that there
was no error being handled for a positive but missing area's Id number.
This would result in a nullptr being dereferenced which is a significant
(seg. faulting, i.e. fatal to the application) issue!

Looking more closely I can see that in (my) PR #4604 a couple of months
back I put a `QMap` container into every area so that all of them may
contain map labels, whereas the prior code held all the QMaps in another
centralised QMap and the prior code checked for a container for an area
when a label was sought - my revised design was what introduced this
regression.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>