Coverity: fix issue registered as CID 1414989 (x10 ?)
Technically we are storing a pointer in a QMap/QHash type container so as
well as checking that there is an entry for the particular key we also
should check that the stored key is non-null before we use it in further
checks.

Whilst we could do this by converting the existing:
`if (originalExits.contains(dirCode)
     && originalExits.value(dirCode)->destination > 0) {`
to be:
`if (originalExits.contains(dirCode)
     && originalExits.value(dirCode)
     && originalExits.value(dirCode)->destination > 0) {`
we can just as well rely on the QMap/QHash::value() method to return a
default constructed (i.e. `nullptr`) if the key provided is not found:
`if (originalExits.value(dirCode)
     && originalExits.value(dirCode)->destination > 0) {`

Further inspection shows that a more compact modification for the LAST exit
being considered in that chunk of code (the one not in the 11 instances
picked up by Coverity) has already been done so it can be applied to these
as well (with a renaming of the local variable from `exit` to `pExit`)!
`auto pExit = originalExits.value(dirCode);
 if (pExit && pExit->destination > 0) {`

It also seems that the same issue is also present in 12 cases in a higher
`if` that contains each of the 11 (or 12) pieces of code in this issue.

This commit should silence this warning in this file but I think it also
appears in other parts of our code (we store a pointer in a container and
check for the presence of the pointer but not that it is not a null one
before trying to dereference it)...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>