Coverity: fix issue registered as CID 1415026 (x6 ?)
In - it seems 6 places we allocate value to a `TExit` (class is in same
file as `dlgRoomExit` class) without checking that the `TExit` was `new`ed
correctly and stored into a container class. This PR addresses that by
checking (with a `Q_ASSERT_X`) that the item was instantiated and avoids
the perceived risk of the container that would be used to hold it not
having it (and thus returning a `nullptr` instead) and also by working with
the tested pointer directly. The container has the item added at the end of
the section of code concerned.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>