Revise: some changes to Server GUI downloader handling (CID 1478854)
Initially this was going to address Coverity CID 1478854 - like Vadi's
PR #5130. However I spotted some improvements that could be made as well:
* rename `(void) cTelnet::replyFinished(QNetworkReply*)` to
`(void) cTelnet::slot_replyFinished(QNetworkReply*)` as it *is* a SLOT.
* rename `(QNetworkReply*) cTelnet::packageDownloadReply` to
`(QNetworkReply*) mpPackageDownloadReply`
* check that the argument passed to the slot method matches the one that is
expected (issues a warning if it doesn't)
* clean up the argument supplied to the slot method in ALL cases.
* reset the stored `QNetworkReply` point to a `nullptr` upon success (and
thus returns it to the state it now has on class creation.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>