Enhance: extend getOS() function to include version information
This extra information is enough to enable Lua scripts to distinguish
between Windows 10 and Windows 7 (the original intention) but during
I found that returning a version number as a second return value is not
itself helpful for GNU/Linux distributions - e.g. for my main Devuan box
it produced the value "3" (for Devaun Stable or 3.0 "Beowuld") so for
Linux only this was extended even further to include the "product name"
(typically a string relating to the "Distribution") as a third return
value.

Note that the Qt functions/macros used are not sufficient (AFAICT) to
report the precise version/build for Windows 10 - which sadly was wanted by
one person concerned about gaining this information somehow.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>