Coverity: fix issue registered as CID 1503736 (#5137)
The original moan was about a check for the `(TMap*) T2DMap::mpMap` member
being a nullptr when it had already been dereferenced several (many) times
before in the same method:
`(void) T2DMap::mouseDoubleClickEvent(QMouseEvent*)`
to correct this I moved the test to the top of method - and I spotted that
some of the related methods had a similar check and others did not. So
I put a similar check in all of them.

Then studying the code more closely I spotted that the class had both a
`(QPointer<Host>) T2DMap::mpHost` and a `(TMap*) T2DMap::mpMap` yet in
quite a few places was using the pointed to class instances own pointer to
the other class instance. This looks to be doing unnecessary indirection
so I have changed them to use this own class's member values as the
pointers instead. Indeed in a further case involving I think some update
code copy-pasted from elsewhere I found some indirection that would in fact
resolve to point to the class `T2Map` instance directly!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>