Bodge: prevent BEL and CR characters from causing display problems
This also intercepts all other control code except TAB and LF and replaces
them with the corresponding Unicode control picture code-point from the
Unicode "control pictures" range (U+2400 to U+2421).

It also add a system "beep" to the handler for the BEL code that is
processed and previously just caused an application "alert" - though
because of the *way* it is processed it will not produce any sound if
`feedTriggers(...)` includes that character - only if it is in the data
sent from Game Server.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>