Enhance: add Push-down Button state getters and setters (Button Luv Part 1)
Provide a means to get and set the state of push down buttons (the latter
without executing the associate command or script). It requires adding
members to the `TAction` class to keep track of the `EAction` or
`TFlipButton` instance that is the implimentation of the button on a
`TEasyButton` or `TToolBar` respectively. As it happens this showed that
the previous code to regenerate them was creating new buttons each time but
was not deleting the existing ones that were getting replaced.

Also:
* made some members of the `TAtion` class (the name, script and both
commands) private as otherwise there is no point in the setters and getters
also provided!
* simpliedied the existing `getButtonState()` {that only works within the
script for a button) to remove an unneeded intermediate local (`int`)
value.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>