BugFix: repair loss of custom profile icons
This should close #5248 which looks to have been introduced by my #3948
from last July. It failed to accomodate the custom icon code from a couple
of years ago done by #2814.

Basically the problem was that I removed storage of the profile name from
the "display text" for the `QTreeWidgetItem`s in the profile connection
dialogue. It was moved to a UserRole stored in the item's data area so that
it was stored but not shown under the item and just showed the icon.
However the custom icon code was still looking at the item's `text()` which
now, of course, is empty!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>