Temporary: combine 64Bit Windows branch with UTF8 Fix attempt [skip gha]
This is a testing branch that is NOT intended to ever be merged - it is
just used to try and compile a pure MSYS2/Mingw-w64 build with code that
reports some text encoder usage as part of the attempt to solve #5262.

As Leris/Kebap found in issue #5262 Windows in particular does not default
to using UTF-8 when reading and writing text files - instead it can use an
8-bit local which (probably) depends on the end-users OS country settings.
However this is not UTF-8 and that will cause issues if such files are
used on another system or OS that does not use the same setting.

This commit forces all uses of `QTextStream` (that are not associated with
a `QString` as the underlying "device" for which it is not applicable) to
use the UTF-8 encoding.

However retro-fitting this change *may* (checks will be useful) mess with
existing files, specifically:
* The config.lua file that contains details of packages/modules in the
  `.mpackage` ZIP archive file format. This is the bug that this is
  intended to cure!
* Reading of the external Lua files that Mudlet uses itself, the
  `LuaGlobal.lua` file and the others that it itself loads - however since
  those are pure ASCII there should not be an issue with them.
* (WIndows ONLY) Reading the external `utf8_filenames.lua` file that is
  used to patch the Lua IO handlers to work with the UTF-16 file names that
  Windows uses in non-en locales. There shouldn't be an issue with this.
* Custom user dictionaries - it is possible that the existing code was
  making use of UTF-8 anyhow - it needs someone like Leris/Kebap to check
  that any custom dictionaries still work given that they had a system
  this encoding problem was happening...!

* report of name of QTextCodec actually used to write config.lua

This is to test what is used on @Kebap (Leris) system which doesn't seem to
work with the "fixes" this PR is supposed to introduce...

* add an additional informational message about QTextCodec selection

This will report whether a codec was actually generated (as `pCodec` will
not be a `nullptr` in that case) - and then report the name it has. This
should clarify whether one was correctly generated for UTF-8 as opposed to
the previous one that indicated that a local8Bit one was being used in some
cases on Windows (as it was called "system") instead of the expected UTF-8.

Signed-off-by: Stephen Lyons <stephen@Hunt>