Enhance: change exit status indication in room exits dialogue
It has been pointed out that using colours to indicate whether an exit room
id number is valid or not in the room exits dialogue is not always helpful
when the end-user wants to apply styling to Mudlet - or for those with some
visual impairments.  This commit is intended to improve the situation by
instead applying an icon to indicate the exit status with one of three
icons indicating an invalid exit room id number, a valid number in the
same area or a valid number in another area. This is shown as an icon in
the `QLineEdit` on the right side (after the number) for *normal* exits and
as an icon in a separate column to the right of the field used to enter and
display the exit room id for *special* exits. This replaces the previous
code that only showed valid or invalid numbers with blue or red text
respectively.

As these changes needed some reorganisation of the individual columns used
for the special exits in the room exit dialogue I converted all the column
numbers into constants that are defined at the top of the `dlgRoomExit.cpp`
file - this means it is more easy to readjust them in the future if that
should prove necessary (and reduces the chance of missing out or mixing up
those index numbers in the event of further editing...!)

During development I discovered how to use custom *item delegates*, e.g.
like described in:
  https://doc.qt.io/qt-5/model-view-programming.html#a-simple-delegate

for both:
* the special exit weight - during editing the `QLineEdit` that was used
  for special exit weights is replaced with a `QSpinBox` that acts in the
  same manner as those for the normal exits.
* the special exit room ID -  during editing a replacement `QLineEdit`
  is used which gains an icon in the same manner as the ones used for the
  normal exits. Outside of that editing a separate column holds the same
  icon except that it also considers the status of the exit command/script
  so as to show an "invalid" exit if there isn't one. Tooltips for both the
  exit room ID and status columns are also updated during and after editing
  of the exit room ID.

Also:
* refactored the initialisation process for the room exits dialogue so that
  it is no longer necessary to explicitly call an `init(...)` method after
  instantiating the dialogue - that is now done from the constructor
  itself - which takes the additional argument (the room that the exits are
  to be displayed for) that was originally needed from that method.
* reordered the items in the `QGridLayout`s in the `room_exits.ui` form XML
  file so that they are back in ascending order again.
* refactored out some highly repetitive (one for each normal exit
  direction) code to some separate methods that take pointers to all the
  individual widgets involved in that exit:
    * `normalExitEdited(...)`
    * `normalStubExitChanged(...)`
* similarly for special exits:
    * `setIconAndToolTipsOnSpecialExit(QTreeWidgetItem*)`
* also for the tooltips for normal exits - which now also include the area
  involved for exits that go to other areas:
    * `generateToolTip(...)`
* and to put an icon (as an `QAction` with no action) into the
  `QLineEdit`s for normal exits:
    * `setActionOnExit(QLineEdit*, QAction*)`
* refactored out some translated `QString`s that are used programmatically
  for special exits - so that they are only specified in one place -
  thereby making the code more robust so that each usage can be revised by
  only editing one string instead of six or seven instances in one case.
* added in a slot method connected to the special exits'
  `QTreeWidget::itemChanged` signal so that it updates better during
  editing of special exits (after moving to or from the exit room id or
  command fields to others within the same special exit instead of only
  when moving to a *different* special exit completely).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>